refactor: split FilesystemTools into Read/Write/Search classes (#115) (#115)
* refactor: split FilesystemTools into Read/Write/Search classes (#116)

* refactor: address AI review feedback on filesystem split

- Add explicit FileNotFoundError handling in read_file, write_file, overwrite_file, and edit_file
- Catches race conditions where file is deleted between exists() check and os.open() call
- Provides clear user-facing error messages with ls() hint for file discovery
- No behavioral changes to existing error flows; FileNotFoundError was previously caught by the broader OSError handler

* docs: update handoff — MR #15 merged, MR #13 in review with fixes

* test: add race condition tests for FileNotFoundError handling

- Add 6 test cases covering FileNotFoundError race conditions in read_file, write_file, overwrite_file, edit_file, and file_info
- Uses unittest.mock.patch to simulate file deletion between exists() check and os.open() call
- Verifies clear user-facing error messages with ls() hint for file discovery
- Includes workspace name hint test for read_file
- All tests pass: 2,840 total, ruff clean

* refactor: address AI review feedback — sanitize errors, limit cap, regex validation

- Sanitize OSError messages in write_file, overwrite_file, edit_file
  to prevent internal path/system details from leaking to users
- Add _MAX_READ_LINES (5,000) cap to read_file limit parameter
  to prevent accidental performance issues with large defaults
- Add regex validation to grep_files before calling search backends
  catches malformed patterns early with clear error messages
- Add 4 new tests covering: limit cap, sanitized errors, regex validation
- All 2,844 tests pass, ruff clean

---------
